build{_daemon,-packages}: stop using inotify; pass package list on argv.
Using inotify(wait) made the code awful, unreliable, slow and with calls to
"sleep 4". Sure it was shiny asynchronous stuff. Sure it was too
complicated.

Send the data as before but use argv to give the list of packages to build
instead of letting build_daemon guess it with inotifywait.

And while at it, avoid spawning some binaries in build-packages.sh to make
the script faster (from 1.35s to 0.25s for a no-op).