build-packages: fix the check dictating whether a package should be built.
If you asked to build the package named "a", the previous test would select
all packages with "a" in the name.
While this can be taken advantage of, it also lead to some false positives
because there are packages which name is a subset of the name of other
packages.