build-packages.sh: assimilate release.sh.
release.sh copied logs, packages and ran sherpa_gen on the package
directories. It did that for every series and was only run at the end of
build-packages.sh.
This means it took a long time to have the log files copied and it also
seemed a little out of place.

Instead, do the copies and sherpa_gen right from build-packages.sh when a
build series is finished building.