build_packages: migrate build infrastructure to OCaml.
This completely removes the old build infrastructure which was written in
shell scripts and replaces it with one written in OCaml.

The main advantage is the ability to schedule builds and to build for both
architectures at the same time, greatly reducing build times.

There are still a few things to refactor and clean but everything works.