chroot.sh, config, hall_of_shame: all deprecated, remove.
The chroot is not the default way to run win-builds nor build the packages.
Chroots will be provided but it will be the responsibility of the user to
setup and enter them which is more flexible and simpler. Moreover, the main
use of chroot.sh was to start build_daemon which has been removed a few
weeks ago.

config was used only by build_daemon which has been removed.

hall-of-shame was a try at documenting the issues seen while packaging so
they could be solved. And let some steam go by ranting and raging a bit.
This would only antagonize people so remove it.