Build/Config: don't make "slack-desc" a dependency of a package.
It causes a few issues and is probably not worth it.
The main issue is that the file might be named differently and it is for
gcc (there are several description files, none of them is named
"slack-desc"). Simply require the dev to remove the .txz or touch some
dependency in order to rebuild the package.