Build/Build: again compute packages to build on the fly.
The list of packages to build used to be computed at the start of the
builder because otherwise a package like gcc:core would make gcc:full
up-to-date since the filenames are the same.

Now that build is negative for the "core" version of these packages, this
is not needed anymore.

This will also avoid failing in the case described above when build is
interrupted after the core version has built and the build is restarted.