build: split the definition of builders from the additions of packages.
In the definition of builder, the windows series depends on the cross-
toolchain one which in turn depends on the native-toolchain series.
However this is reversed for the package dependencies: the packages from
the windows series can have dependencies on the packages from the cross-
toolchain series.

This split makes it possible to add cross-series dependencies.